Move resources into its own package.
