+++ resources_test.go
@@ -85,324 +85,3 @@ func (s *resourceSuite) TestSchemaMissingComment(c *gc.C) {
 		"comment":  "",
 	})
 }
-
-func (s *resourceSuite) TestParseResourceTypeOkay(c *gc.C) {
-	rt, ok := charm.ParseResourceType("file")
-
-	c.Check(ok, jc.IsTrue)
-	c.Check(rt, gc.Equals, charm.ResourceTypeFile)
-}
-
-func (s *resourceSuite) TestParseResourceTypeRecognized(c *gc.C) {
-	supported := []charm.ResourceType{
-		charm.ResourceTypeFile,
-	}
-	for _, expected := range supported {
-		rt, ok := charm.ParseResourceType(expected.String())
-
-		c.Check(ok, jc.IsTrue)
-		c.Check(rt, gc.Equals, expected)
-	}
-}
-
-func (s *resourceSuite) TestParseResourceTypeEmpty(c *gc.C) {
-	rt, ok := charm.ParseResourceType("")
-
-	c.Check(ok, jc.IsFalse)
-	c.Check(rt, gc.Equals, charm.ResourceTypeUnknown)
-}
-
-func (s *resourceSuite) TestParseResourceTypeUnsupported(c *gc.C) {
-	rt, ok := charm.ParseResourceType("spam")
-
-	c.Check(ok, jc.IsFalse)
-	c.Check(rt, gc.Equals, charm.ResourceType("spam"))
-}
-
-func (s *resourceSuite) TestResourceTypeStringSupported(c *gc.C) {
-	supported := map[charm.ResourceType]string{
-		charm.ResourceTypeFile: "file",
-	}
-	for rt, expected := range supported {
-		str := rt.String()
-
-		c.Check(str, gc.Equals, expected)
-	}
-}
-
-func (s *resourceSuite) TestResourceTypeStringUnknown(c *gc.C) {
-	str := charm.ResourceTypeUnknown.String()
-
-	c.Check(str, gc.Equals, "<unknown>")
-}
-
-func (s *resourceSuite) TestResourceTypeStringUnsupported(c *gc.C) {
-	str := charm.ResourceType("spam").String()
-
-	c.Check(str, gc.Equals, "spam")
-}
-
-func (s *resourceSuite) TestResourceTypeValidateSupported(c *gc.C) {
-	supported := []charm.ResourceType{
-		charm.ResourceTypeFile,
-	}
-	for _, rt := range supported {
-		err := rt.Validate()
-
-		c.Check(err, jc.ErrorIsNil)
-	}
-}
-
-func (s *resourceSuite) TestResourceTypeValidateUnknown(c *gc.C) {
-	err := charm.ResourceTypeUnknown.Validate()
-
-	c.Check(err, gc.ErrorMatches, `unsupported resource type .*`)
-}
-
-func (s *resourceSuite) TestResourceTypeValidateUnsupported(c *gc.C) {
-	err := charm.ResourceType("spam").Validate()
-
-	c.Check(err, gc.ErrorMatches, `unsupported resource type .*`)
-}
-
-func (s *resourceSuite) TestParseResourceOkay(c *gc.C) {
-	name := "my-resource"
-	data := map[string]interface{}{
-		"type":     "file",
-		"filename": "filename.tgz",
-		"comment":  "One line that is useful when operators need to push it.",
-	}
-	resource := charm.ParseResource(name, data)
-
-	c.Check(resource, jc.DeepEquals, charm.Resource{
-		ResourceInfo: charm.ResourceInfo{
-			Name:    "my-resource",
-			Type:    charm.ResourceTypeFile,
-			Path:    "filename.tgz",
-			Comment: "One line that is useful when operators need to push it.",
-		},
-	})
-}
-
-func (s *resourceSuite) TestParseResourceMissingName(c *gc.C) {
-	name := ""
-	data := map[string]interface{}{
-		"type":     "file",
-		"filename": "filename.tgz",
-		"comment":  "One line that is useful when operators need to push it.",
-	}
-	resource := charm.ParseResource(name, data)
-
-	c.Check(resource, jc.DeepEquals, charm.Resource{
-		ResourceInfo: charm.ResourceInfo{
-			Name:    "",
-			Type:    charm.ResourceTypeFile,
-			Path:    "filename.tgz",
-			Comment: "One line that is useful when operators need to push it.",
-		},
-	})
-}
-
-func (s *resourceSuite) TestParseResourceMissingType(c *gc.C) {
-	name := "my-resource"
-	data := map[string]interface{}{
-		"filename": "filename.tgz",
-		"comment":  "One line that is useful when operators need to push it.",
-	}
-	resource := charm.ParseResource(name, data)
-
-	c.Check(resource, jc.DeepEquals, charm.Resource{
-		ResourceInfo: charm.ResourceInfo{
-			Name:    "my-resource",
-			Type:    charm.ResourceTypeUnknown,
-			Path:    "filename.tgz",
-			Comment: "One line that is useful when operators need to push it.",
-		},
-	})
-}
-
-func (s *resourceSuite) TestParseResourceMissingPath(c *gc.C) {
-	name := "my-resource"
-	data := map[string]interface{}{
-		"type":    "file",
-		"comment": "One line that is useful when operators need to push it.",
-	}
-	resource := charm.ParseResource(name, data)
-
-	c.Check(resource, jc.DeepEquals, charm.Resource{
-		ResourceInfo: charm.ResourceInfo{
-			Name:    "my-resource",
-			Type:    charm.ResourceTypeFile,
-			Path:    "",
-			Comment: "One line that is useful when operators need to push it.",
-		},
-	})
-}
-
-func (s *resourceSuite) TestParseResourceMissingComment(c *gc.C) {
-	name := "my-resource"
-	data := map[string]interface{}{
-		"type":     "file",
-		"filename": "filename.tgz",
-	}
-	resource := charm.ParseResource(name, data)
-
-	c.Check(resource, jc.DeepEquals, charm.Resource{
-		ResourceInfo: charm.ResourceInfo{
-			Name:    "my-resource",
-			Type:    charm.ResourceTypeFile,
-			Path:    "filename.tgz",
-			Comment: "",
-		},
-	})
-}
-
-func (s *resourceSuite) TestParseResourceEmpty(c *gc.C) {
-	name := "my-resource"
-	data := make(map[string]interface{})
-	resource := charm.ParseResource(name, data)
-
-	c.Check(resource, jc.DeepEquals, charm.Resource{
-		ResourceInfo: charm.ResourceInfo{
-			Name: "my-resource",
-		},
-	})
-}
-
-func (s *resourceSuite) TestParseResourceNil(c *gc.C) {
-	name := "my-resource"
-	var data map[string]interface{}
-	resource := charm.ParseResource(name, data)
-
-	c.Check(resource, jc.DeepEquals, charm.Resource{
-		ResourceInfo: charm.ResourceInfo{
-			Name: "my-resource",
-		},
-	})
-}
-
-func (s *resourceSuite) TestValidateFull(c *gc.C) {
-	resource := charm.Resource{
-		ResourceInfo: charm.ResourceInfo{
-			Name:    "my-resource",
-			Type:    charm.ResourceTypeFile,
-			Path:    "filename.tgz",
-			Comment: "One line that is useful when operators need to push it.",
-		},
-	}
-	err := resource.Validate()
-
-	c.Check(err, jc.ErrorIsNil)
-}
-
-func (s *resourceSuite) TestValidateZeroValue(c *gc.C) {
-	var resource charm.Resource
-	err := resource.Validate()
-
-	c.Check(err, gc.NotNil)
-}
-
-func (s *resourceSuite) TestValidateMissingName(c *gc.C) {
-	resource := charm.Resource{
-		ResourceInfo: charm.ResourceInfo{
-			Type:    charm.ResourceTypeFile,
-			Path:    "filename.tgz",
-			Comment: "One line that is useful when operators need to push it.",
-		},
-	}
-	err := resource.Validate()
-
-	c.Check(err, gc.ErrorMatches, `resource missing name`)
-}
-
-func (s *resourceSuite) TestValidateMissingType(c *gc.C) {
-	resource := charm.Resource{
-		ResourceInfo: charm.ResourceInfo{
-			Name:    "my-resource",
-			Path:    "filename.tgz",
-			Comment: "One line that is useful when operators need to push it.",
-		},
-	}
-	err := resource.Validate()
-
-	c.Check(err, gc.ErrorMatches, `resource missing type`)
-}
-
-func (s *resourceSuite) TestValidateUnknownType(c *gc.C) {
-	resource := charm.Resource{
-		ResourceInfo: charm.ResourceInfo{
-			Name:    "my-resource",
-			Type:    "repo",
-			Path:    "repo-root",
-			Comment: "One line that is useful when operators need to push it.",
-		},
-	}
-	err := resource.Validate()
-
-	c.Check(err, gc.ErrorMatches, `.*unsupported resource type .*`)
-}
-
-func (s *resourceSuite) TestValidateMissingPath(c *gc.C) {
-	resource := charm.Resource{
-		ResourceInfo: charm.ResourceInfo{
-			Name:    "my-resource",
-			Type:    charm.ResourceTypeFile,
-			Comment: "One line that is useful when operators need to push it.",
-		},
-	}
-	err := resource.Validate()
-
-	c.Check(err, gc.ErrorMatches, `resource missing filename`)
-}
-
-func (s *resourceSuite) TestValidateNestedPath(c *gc.C) {
-	resource := charm.Resource{
-		ResourceInfo: charm.ResourceInfo{
-			Name: "my-resource",
-			Type: charm.ResourceTypeFile,
-			Path: "spam/eggs",
-		},
-	}
-	err := resource.Validate()
-
-	c.Check(err, gc.ErrorMatches, `.*filename cannot contain "/" .*`)
-}
-
-func (s *resourceSuite) TestValidateAbsolutePath(c *gc.C) {
-	resource := charm.Resource{
-		ResourceInfo: charm.ResourceInfo{
-			Name: "my-resource",
-			Type: charm.ResourceTypeFile,
-			Path: "/spam/eggs",
-		},
-	}
-	err := resource.Validate()
-
-	c.Check(err, gc.ErrorMatches, `.*filename cannot contain "/" .*`)
-}
-
-func (s *resourceSuite) TestValidateSuspectPath(c *gc.C) {
-	resource := charm.Resource{
-		ResourceInfo: charm.ResourceInfo{
-			Name: "my-resource",
-			Type: charm.ResourceTypeFile,
-			Path: "git@github.com:juju/juju.git",
-		},
-	}
-	err := resource.Validate()
-
-	c.Check(err, gc.ErrorMatches, `.*filename cannot contain "/" .*`)
-}
-
-func (s *resourceSuite) TestValidateMissingComment(c *gc.C) {
-	resource := charm.Resource{
-		ResourceInfo: charm.ResourceInfo{
-			Name: "my-resource",
-			Type: charm.ResourceTypeFile,
-			Path: "filename.tgz",
-		},
-	}
-	err := resource.Validate()
-
-	c.Check(err, jc.ErrorIsNil)
-}
